+++ src/Common/Class/GameScene/PhysicEngine/Speed/query.cpp
@@ -1,11 +1,11 @@
 # include "gameScene.h"
 
-int     CS_Speed::QuerySpeedX()
+float     CS_Speed::QuerySpeedX()
 {
     return (vX);
 }
 
-int     CS_Speed::QuerySpeedY()
+float     CS_Speed::QuerySpeedY()
 {
     return (vY);
 }
